Fixes for windows and @ path naming
On Windows (at least in git-for-windows bash shell) the yarn test
target needed the paths not to be quoted. Worked on Mac but it's
better this way anyway, as quotes are needed in shells, not JSON ;)

Also added the tsconfig changes to let the `@/...` syntax work
to name the source dir, which was only needed for the tests/**/*.ts
files for some reason.